Pull CollectionChangeBuilder into its own file
