+++ Realm/ObjectStore/impl/collection_notifier.hpp
@@ -19,7 +19,7 @@
 #ifndef REALM_BACKGROUND_COLLECTION_HPP
 #define REALM_BACKGROUND_COLLECTION_HPP
 
-#include "collection_notifications.hpp"
+#include "impl/collection_change_builder.hpp"
 
 #include <realm/group_shared.hpp>
 
@@ -34,41 +34,6 @@ namespace realm {
 class Realm;
 
 namespace _impl {
-class CollectionChangeBuilder : public CollectionChangeSet {
-public:
-    CollectionChangeBuilder(CollectionChangeBuilder const&) = default;
-    CollectionChangeBuilder(CollectionChangeBuilder&&) = default;
-    CollectionChangeBuilder& operator=(CollectionChangeBuilder const&) = default;
-    CollectionChangeBuilder& operator=(CollectionChangeBuilder&&) = default;
-
-    CollectionChangeBuilder(IndexSet deletions = {},
-                            IndexSet insertions = {},
-                            IndexSet modification = {},
-                            std::vector<Move> moves = {});
-
-    static CollectionChangeBuilder calculate(std::vector<size_t> const& old_rows,
-                                             std::vector<size_t> const& new_rows,
-                                             std::function<bool (size_t)> row_did_change,
-                                             bool sort);
-
-    void merge(CollectionChangeBuilder&&);
-    void clean_up_stale_moves();
-
-    void insert(size_t ndx, size_t count=1, bool track_moves=true);
-    void modify(size_t ndx);
-    void erase(size_t ndx);
-    void move_over(size_t ndx, size_t last_ndx, bool track_moves=true);
-    void clear(size_t old_size);
-    void move(size_t from, size_t to);
-
-    void parse_complete();
-
-private:
-    std::unordered_map<size_t, size_t> m_move_mapping;
-
-    void verify();
-
-};
 struct ListChangeInfo {
     size_t table_ndx;
     size_t row_ndx;
